assign-2/requirements.txt
pytz
python-dateutil
cython
pycocotools
opencv-python
imutils
numpy
pandas
einops
Pillow
requests